Move Critique Clients to their own folder
