+++ src/helm/proxy/critique/mechanical_turk_critique_client.py
@@ -2,9 +2,9 @@ from helm.common.critique_request import (
     CritiqueRequest,
     CritiqueRequestResult,
 )
-from helm.proxy.clients.critique_client import CritiqueClient
-from helm.proxy.clients.mechanical_turk_critique_exporter import export_request
-from helm.proxy.clients.mechanical_turk_critique_importer import import_request_result
+from helm.proxy.critique.critique_client import CritiqueClient
+from helm.proxy.critique.mechanical_turk_critique_exporter import export_request
+from helm.proxy.critique.mechanical_turk_critique_importer import import_request_result
 
 
 class MechanicalTurkCritiqueClient(CritiqueClient):
