+++ src/helm/proxy/critique/scale_critique_client.py
@@ -15,7 +15,7 @@ from helm.common.critique_request import (
     CritiqueResponse,
 )
 from helm.common.optional_dependencies import handle_module_not_found_error
-from helm.proxy.clients.critique_client import CritiqueClient
+from helm.proxy.critique.critique_client import CritiqueClient
 
 try:
     import scaleapi
